requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
appdirs==1.4.3
args==0.1.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
clint==0.5.1
cmake==3.30.5
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.0
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==1.3.5
pillow==10.4.0
pip==24.3.1
plyfile==1.0.3
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-tgb==0.1.1
pyg-lib==0.4.0+pt20cu117
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
setuptools==75.3.0
six==1.14.0
sympy==1.13.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.0.0
torch-cluster==1.6.3+pt20cu117
torch-geometric==2.6.1
torch-scatter==2.1.2+pt20cu117
torch-sparse==0.6.18+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
torch_geometric_temporal==0.54.0
torchsummary==1.5.1
tqdm==4.66.6
transformers==4.38.2
triton==2.0.0
typing_extensions==4.7.1
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.0.6
wheel==0.44.0
yarl==1.15.2
zipp==3.20.2